third_party/blink/web_tests/webaudio/DynamicsCompressor/dynamicscompressor-clear-internal-state.html

<!DOCTYPE html>
<html>
  <head>
    <title>
      Validate Reduction Value of DynamicsComporessor after Disabling
    </title>
    <script src="../../resources/testharness.js"></script>
    <script src="../../resources/testharnessreport.js"></script>
    <script src="../resources/audit-util.js"></script>
  </head>
  <body>
    <script>

      /*
       * - Feed a DC signal (value = 1, duration = 5 s)
           through a DynamicsCompressor.
       * - Render 10 s total; after the DC ends the compressor should release
       *   and its |reduction| should converge → 0.
       * - Verify |reduction| ≤ 4.8223 e-2.
       */

      promise_test(t => {
      // Use a low sample rate to reduce complexity because we need to run for
      // quite a few seconds to get the reduction to converge.
        const sampleRate = 8192;
        const testDurationSamples = 10 * sampleRate;
        // Create the offline context for the test.
        const context = new OfflineAudioContext(
            /* channels */ 1,
            /* frames   */ testDurationSamples,
            /* rate     */ sampleRate);

        // 5 s buffer full-scale DC (value = 1)
        const buffer = createConstantBuffer(
            context, testDurationSamples / 2, 1);

        // Create the source and connect it to the destination
        const source = new AudioBufferSourceNode(context);

        // Create compressor and use default parameters for the compression.
        const compressor = new DynamicsCompressorNode(context);

        source.buffer = buffer;
        source.connect(compressor);
        compressor.connect(context.destination);
        source.start(0.0);

        // Error threshold experimentally determined.
        const ERROR_THRESHOLD = 4.8223e-2;

        // Return the promise so testharness waits on it.
        return context.startRendering().then(() => {
          const reduction = Math.abs(compressor.reduction);

          assert_less_than_equal(
              reduction, ERROR_THRESHOLD,
              `|reduction| (${reduction}) should be ≤ ${ERROR_THRESHOLD}`);
        });
      }, 'DynamicsCompressor.reduction approaches 0 after long silence');
    </script>
  </body>
</html>
